Make ApiClient.request public; support sdp request & text response
The purpose of this is to support writing a `RealtimeClient` class in a separate module.
I do not want that code to have to rewrite request/response serialization.
Conversely, I would like for the code to reuse as much of this module as useful.
That code will be serializing RealtimeClientEvent* requests a lot.
I see no reason why ApiClient.request **has** to be protected.

The problem is that ApiClient.request is inline reified, and it seems that all inline
reified methods need to have all methods they call have the same visibility. :/
It is definitely not ideal that I am making so many protected things public, but
the methods I am making public seem harmless enough to do coding a more
drastic refactor.